feat(dracut): create sysusers only when the initrd is generated
When generating the initrd, dracut copies all the sysusers dropins provided by
its modules to it, and systemd-sysusers runs at every boot to always create the
same users (`ConditionNeedsUpdate` removed in #1658).

This patch is intended to create the system users only when the initrd is
generated, without including the systemd-sysusers stuff in the initrd, as it is
not necessary.

To control this functionality and allow to use the old method, added the
`create_sysusers` configuration option (default to yes) and the
`--no-create-sysusers` parameter.