feat(dmsquash-live): add support for rd.live.overlay.nouserconfirmprompt
Dracut allows the creation of a LiveOS using a read-only squashfs and
a read-write overlay on top.

If the read-write overlay is backed by a ramdisk, Dracut halts and prompts
the user to confirm whether to continue or not.

This interaction during the boot process is not desired in all cases.

This change introduces a new flag (rd.live.overlay.nouserconfirmprompt)
that when defined, it suppresses the prompt and allows the boot process
to continue to completion without user interation.

There is no impact to existing configurations and their associated behavior.
The new behavior only takes effect only when the new switch is explicitly
defined by the image build as a kernel parameter.

(cherry picked from commit https://github.com/dracut-ng/dracut-ng/commit/6ac1033c8d59e2324f56a2c453f14a0318cc7190)