feat(dracut): take advantage of the new /sys/firmware/initrd sysfs file
Since v6.8-rc1 [1], when the kernel was booted with an initrd and the
`retain_initrd` option is set on the kernel command line, /sys/firmware/initrd
contains the contents of the initrd that the kernel was booted with.

Therefore, there is no need to guess the path of the initrd for `lsinitrd`
and `dracut-initramfs-restore` if this sysfs file exists on the running system.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=2678fd2fe9ee2c569e9cb6b17e786bc8f0753538